+++ CHANGELOG.md
@@ -4,6 +4,7 @@ All notable changes to this project will be documented in this file.
 ## [Unreleased]
 ### Added
 - Endpoint returning last transaction reference for a given address
+- Additional validation checking last transaction reference when submitting a transaction
 
 ## [v2.6.0] - 2020-05-08
 ### Changed